+++ src/main_nep_adam/dataset.cuh
@@ -28,7 +28,7 @@ public:
   int max_Na;         // number of atoms in the largest configuration
   int max_NN_radial;  // radial neighbor list size
   int max_NN_angular; // angular neighbor list size
-  double sum_energy_ref;                  // sum of reference energy for Nc
+  float sum_energy_ref;                  // sum of reference energy for Nc
   bool all_type;  // whether include all types for structures
 
 
@@ -40,37 +40,37 @@ public:
 
   GPU_Vector<int> type;           // atom type (0, 1, 2, 3, ...)
   GPU_Vector<int> type_sum;      // prefix sum of type
-  GPU_Vector<double> r;            // position
-  GPU_Vector<double> box;          // (expanded) box and inverse box (18 components)
-  GPU_Vector<double> box_original; // (original) box (9 components)
+  GPU_Vector<float> r;            // position
+  GPU_Vector<float> box;          // (expanded) box and inverse box (18 components)
+  GPU_Vector<float> box_original; // (original) box (9 components)
   GPU_Vector<int> num_cell;       // number of cells in the expanded box (3 components)
 
-  GPU_Vector<double> energy;      // calculated energy in GPU
-  GPU_Vector<double> virial;      // calculated virial in GPU
-  GPU_Vector<double> force;       // calculated force in GPU
-  std::vector<double> energy_cpu; // calculated energy in CPU
-  std::vector<double> virial_cpu; // calculated virial in CPU
-  std::vector<double> force_cpu;  // calculated force in CPU
+  GPU_Vector<float> energy;      // calculated energy in GPU
+  GPU_Vector<float> virial;      // calculated virial in GPU
+  GPU_Vector<float> force;       // calculated force in GPU
+  std::vector<float> energy_cpu; // calculated energy in CPU
+  std::vector<float> virial_cpu; // calculated virial in CPU
+  std::vector<float> force_cpu;  // calculated force in CPU
 
-  GPU_Vector<double> energy_ref_gpu;       // reference energy in GPU
-  GPU_Vector<double> virial_ref_gpu;       // reference virial in GPU
-  GPU_Vector<double> force_ref_gpu;        // reference force in GPU
-  GPU_Vector<double> temperature_ref_gpu;  // reference temperature in GPU
-  std::vector<double> energy_ref_cpu;      // reference energy in CPU
-  std::vector<double> virial_ref_cpu;      // reference virial in CPU
-  std::vector<double> force_ref_cpu;       // reference force in CPU
-  std::vector<double> weight_cpu;          // configuration weight in CPU
-  GPU_Vector<double> weight_gpu;          // configuration weight in GPU
-  std::vector<double> temperature_ref_cpu; // reference temeprature in CPU
+  GPU_Vector<float> energy_ref_gpu;       // reference energy in GPU
+  GPU_Vector<float> virial_ref_gpu;       // reference virial in GPU
+  GPU_Vector<float> force_ref_gpu;        // reference force in GPU
+  GPU_Vector<float> temperature_ref_gpu;  // reference temperature in GPU
+  std::vector<float> energy_ref_cpu;      // reference energy in CPU
+  std::vector<float> virial_ref_cpu;      // reference virial in CPU
+  std::vector<float> force_ref_cpu;       // reference force in CPU
+  std::vector<float> weight_cpu;          // configuration weight in CPU
+  GPU_Vector<float> weight_gpu;          // configuration weight in GPU
+  std::vector<float> temperature_ref_cpu; // reference temeprature in CPU
 
-  GPU_Vector<double> type_weight_gpu; // relative force weight for different atom types (GPU)
+  GPU_Vector<float> type_weight_gpu; // relative force weight for different atom types (GPU)
 
-  std::vector<double> error_cpu_e; // error in energy (squared)
-  std::vector<double> error_cpu_v; // error in virial (squared)
-  std::vector<double> error_cpu_f; // error in force (squared)
-  GPU_Vector<double> error_gpu;  // error in energy, virial, or force (squared)
-  GPU_Vector<double> diff_gpu_e; // error in energy (before squared)
-  GPU_Vector<double> diff_gpu_v; // error in virial (before squared)
+  std::vector<float> error_cpu_e; // error in energy (squared)
+  std::vector<float> error_cpu_v; // error in virial (squared)
+  std::vector<float> error_cpu_f; // error in force (squared)
+  GPU_Vector<float> error_gpu;  // error in energy, virial, or force (squared)
+  GPU_Vector<float> diff_gpu_e; // error in energy (before squared)
+  GPU_Vector<float> diff_gpu_v; // error in virial (before squared)
   std::vector<bool> has_type;
   std::vector<bool> has_virial;
 
@@ -79,12 +79,12 @@ public:
 
   void
   construct(Parameters& para, std::vector<Structure>& structures, bool require_grad, int n1, int n2, int device_id);
-  std::vector<double> get_rmse_force(Parameters& para, const bool use_weight, int device_id);
-  std::vector<double> get_rmse_energy(
+  std::vector<float> get_rmse_force(Parameters& para, const bool use_weight, int device_id);
+  std::vector<float> get_rmse_energy(
     Parameters& para,
     const bool use_weight,
     int device_id);
-  std::vector<double> get_rmse_virial(Parameters& para, const bool use_weight, int device_id);
+  std::vector<float> get_rmse_virial(Parameters& para, const bool use_weight, int device_id);
 
 private:
   void copy_structures(std::vector<Structure>& structures_input, int n1, int n2);
